Overhaul how EDS Change Sets are received, processed and observed. [touch:43]
We no longer store a historical reference of all Change Sets that have been received. Once a Change Set is validated and deserialised, observers are responsible for adding queued commands to the queueing system.

From there, no more involvement with an EDS Change Set is undertaken.

All commands that previously referenced a “Change Set ID” (to help with EDS observers) now utilise a “Batch ID” parameter instead. This decouples the commands from EDS and also allows us to provide a Batch ID during regular bulk syncs (for example with products), so that tools such as skipping subsequent syncs of a configurable product can occur on a bulk level as well as in an EDS Change Set.